requirements.txt
Flask
gunicorn
dvc
pandas
notebook
numpy
matplotlib
seaborn
sklearn
kaggle
regex
notebook
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
pathlib
types-PyYAML
evidently
ipykernel
-e .